Refactor showcase examples into modules
